replace dual-backend setup with a single backend
